ajuestes no DAO categoria

+++ produto/carofour/src/com/pucminas/carofour/dao/CategoriaDAO.java
@@ -1,22 +1,17 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 package com.pucminas.carofour.dao;
 
 import com.pucminas.carofour.model.Categoria;
 import com.pucminas.carofour.model.Produto;
+
 import java.util.List;
 
 /**
- * @author jonatas
+ * @version 0.2
  */
 
 public interface CategoriaDAO {
     List<Categoria> listCategorias();
-    Categoria find(Long id);
+    Categoria find(int id);
     boolean save(Categoria categoria);
-    boolean remove(Long id);
+    boolean remove(int id);
 }
